refactor: department 更新スキーマの version を z.coerce.number().int() に統一する (#312)
version 列は Prisma 上 Int（ADR-0039 の楽観ロックトークン）。.int() が無いと
"2.5" のような小数が通過しうるため、整数を強制する role の更新スキーマと表現を揃える。

+++ src/app/(features)/departments/[departmentCd]/schema.ts
@@ -10,7 +10,7 @@ import { departmentBaseSchema } from "../_shared/schema";
  */
 export const updateDepartmentSchema = departmentBaseSchema.extend({
   isActive: z.coerce.boolean(),
-  version: z.coerce.number(),
+  version: z.coerce.number().int(),
 });
 
 export type UpdateDepartmentInput = z.infer<typeof updateDepartmentSchema>;
